cleanup / refactor tests so that they make more sense
