+++ gql-gen/graph/resolver.go
@@ -4,10 +4,10 @@ package graph
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/bmeg/grip-graphql/gql-gen/generated"
 	"github.com/bmeg/grip-graphql/gql-gen/model"
+	"github.com/bmeg/grip/log"
 )
 
 // Organization is the resolver for the organization field.
@@ -19,7 +19,7 @@ func (r *queryResolver) Organization(ctx context.Context, offset *int, first *in
 
 	slice, err := model.UnmarshalOrganizationTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -34,7 +34,7 @@ func (r *queryResolver) Group(ctx context.Context, offset *int, first *int, filt
 
 	slice, err := model.UnmarshalGroupTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -49,7 +49,7 @@ func (r *queryResolver) Practitioner(ctx context.Context, offset *int, first *in
 
 	slice, err := model.UnmarshalPractitionerTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -64,7 +64,7 @@ func (r *queryResolver) PractitionerRole(ctx context.Context, offset *int, first
 
 	slice, err := model.UnmarshalPractitionerRoleTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -79,7 +79,7 @@ func (r *queryResolver) ResearchStudy(ctx context.Context, offset *int, first *i
 
 	slice, err := model.UnmarshalResearchStudyTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -94,7 +94,7 @@ func (r *queryResolver) Patient(ctx context.Context, offset *int, first *int, fi
 
 	slice, err := model.UnmarshalPatientTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -109,7 +109,7 @@ func (r *queryResolver) ResearchSubject(ctx context.Context, offset *int, first 
 
 	slice, err := model.UnmarshalResearchSubjectTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -124,7 +124,7 @@ func (r *queryResolver) Substance(ctx context.Context, offset *int, first *int, 
 
 	slice, err := model.UnmarshalSubstanceTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -139,7 +139,7 @@ func (r *queryResolver) SubstanceDefinition(ctx context.Context, offset *int, fi
 
 	slice, err := model.UnmarshalSubstanceDefinitionTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -154,7 +154,7 @@ func (r *queryResolver) Specimen(ctx context.Context, offset *int, first *int, f
 
 	slice, err := model.UnmarshalSpecimenTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -169,7 +169,7 @@ func (r *queryResolver) Observation(ctx context.Context, offset *int, first *int
 
 	slice, err := model.UnmarshalObservationTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 
@@ -185,7 +185,7 @@ func (r *queryResolver) DiagnosticReport(ctx context.Context, offset *int, first
 
 	slice, err := model.UnmarshalDiagnosticReportTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -200,7 +200,7 @@ func (r *queryResolver) Condition(ctx context.Context, offset *int, first *int, 
 
 	slice, err := model.UnmarshalConditionTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -215,7 +215,7 @@ func (r *queryResolver) Medication(ctx context.Context, offset *int, first *int,
 
 	slice, err := model.UnmarshalMedicationTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -230,7 +230,7 @@ func (r *queryResolver) MedicationAdministration(ctx context.Context, offset *in
 
 	slice, err := model.UnmarshalMedicationAdministrationTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -245,7 +245,7 @@ func (r *queryResolver) MedicationStatement(ctx context.Context, offset *int, fi
 
 	slice, err := model.UnmarshalMedicationStatementTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -260,7 +260,7 @@ func (r *queryResolver) MedicationRequest(ctx context.Context, offset *int, firs
 
 	slice, err := model.UnmarshalMedicationRequestTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -275,7 +275,7 @@ func (r *queryResolver) Procedure(ctx context.Context, offset *int, first *int, 
 
 	slice, err := model.UnmarshalProcedureTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -290,7 +290,7 @@ func (r *queryResolver) DocumentReference(ctx context.Context, offset *int, firs
 
 	slice, err := model.UnmarshalDocumentReferenceTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -305,7 +305,7 @@ func (r *queryResolver) Task(ctx context.Context, offset *int, first *int, filte
 
 	slice, err := model.UnmarshalTaskTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -320,7 +320,7 @@ func (r *queryResolver) ImagingStudy(ctx context.Context, offset *int, first *in
 
 	slice, err := model.UnmarshalImagingStudyTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -335,7 +335,7 @@ func (r *queryResolver) FamilyMemberHistory(ctx context.Context, offset *int, fi
 
 	slice, err := model.UnmarshalFamilyMemberHistoryTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
@@ -350,7 +350,7 @@ func (r *queryResolver) BodyStructure(ctx context.Context, offset *int, first *i
 
 	slice, err := model.UnmarshalBodyStructureTypeSlice(data)
 	if err != nil {
-		fmt.Println("ERR: ", err)
+		log.Errorln("ERR: ", err)
 		return nil, err
 	}
 	return slice, nil
