+++ tests/fixtures/combio-examples-grip/DocumentReference.in.edge.json
@@ -1,2 +1,2 @@
-{"id": "11e86b6c-4e7d-5b94-a892-216e29baf0cb","label": "subject_Specimen","from": "9ae7e542-767f-4b03-a854-7ceed17152cb","to": "60c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id": "01e86b6c-4e7d-5b94-a892-216e29baf0cb","label": "subject_Specimen","from": "8ae7e542-767f-4b03-a854-7ceed17152cb","to": "50c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id": "11e86b6c-4e7d-5b94-a892-216e29baf0cb","_label": "subject_Specimen","_from": "9ae7e542-767f-4b03-a854-7ceed17152cb","_to": "60c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id": "01e86b6c-4e7d-5b94-a892-216e29baf0cb","_label": "subject_Specimen","_from": "8ae7e542-767f-4b03-a854-7ceed17152cb","_to": "50c67a06-ea2d-4d24-9249-418dc77a16a9"}
