+++ tests/fixtures/combio-examples-grip/DocumentReference.out.edge.json
@@ -1,2 +1,2 @@
-{"id": "655b13f1-1955-510f-bac8-5934777099e4","label": "document_reference","from": "60c67a06-ea2d-4d24-9249-418dc77a16a9","to": "9ae7e542-767f-4b03-a854-7ceed17152cb"}
-{"id": "555b13f1-1955-510f-bac8-5934777099e4","label": "document_reference","from": "50c67a06-ea2d-4d24-9249-418dc77a16a9","to": "8ae7e542-767f-4b03-a854-7ceed17152cb"}
+{"_id": "655b13f1-1955-510f-bac8-5934777099e4","_label": "document_reference","_from": "60c67a06-ea2d-4d24-9249-418dc77a16a9","_to": "9ae7e542-767f-4b03-a854-7ceed17152cb"}
+{"_id": "555b13f1-1955-510f-bac8-5934777099e4","_label": "document_reference","_from": "50c67a06-ea2d-4d24-9249-418dc77a16a9","_to": "8ae7e542-767f-4b03-a854-7ceed17152cb"}
