+++ tests/fixtures/combio-examples-grip/DocumentReference.vertex.json
@@ -1,2 +1,2 @@
-{"id":"9ae7e542-767f-4b03-a854-7ceed17152cb","label":"DocumentReference","data":{"auth_resource_path":"/programs/ohsu/projects/test","content":[{"attachment":{"contentType":"text/fastq","creation":"2024-08-21T10:53:00+00:00","extension":[{"url":"http://aced-idp.org/fhir/StructureDefinition/md5","valueString":"227f0a5379362d42eaa1814cfc0101b8"},{"url":"http://aced-idp.org/fhir/StructureDefinition/source_path","valueUrl":"file:///home/LabA/specimen_1234_labA.fq.gz"}],"size":5595609484,"title":"specimen_1234_labA.fq.gz","url":"file:///home/LabA/specimen_1234_labA.fq.gz"}}],"date":"2024-08-21T10:53:00+00:00","docStatus":"final","id":"9ae7e542-767f-4b03-a854-7ceed17152cb","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"9ae7e542-767f-4b03-a854-7ceed17152cb"}],"resourceType":"DocumentReference","status":"current","subject":{"reference":"Specimen/60c67a06-ea2d-4d24-9249-418dc77a16a9"}}}
-{"id":"8ae7e542-767f-4b03-a854-7ceed17152cb","label":"DocumentReference","data":{"auth_resource_path":"/programs/ohsu/projects/test-invalid","content":[{"attachment":{"contentType":"text/fastq","creation":"2024-08-21T10:53:00+00:00","extension":[{"url":"http://aced-idp.org/fhir/StructureDefinition/md5","valueString":"227f0a5379362d42eaa1814cfc0101b8"},{"url":"http://aced-idp.org/fhir/StructureDefinition/source_path","valueUrl":"file:///home/LabA/specimen_1234_labA.fq.gz"}],"size":5595609484,"title":"specimen_1234_labA.fq.gz","url":"file:///home/LabA/specimen_1234_labA.fq.gz"}}],"date":"2024-08-21T10:53:00+00:00","docStatus":"final","id":"8ae7e542-767f-4b03-a854-7ceed17152cb","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"8ae7e542-767f-4b03-a854-7ceed17152cb"}],"resourceType":"DocumentReference","status":"current","subject":{"reference":"Specimen/50c67a06-ea2d-4d24-9249-418dc77a16a9"}}}
+{"_id":"9ae7e542-767f-4b03-a854-7ceed17152cb","_label":"DocumentReference","auth_resource_path":"/programs/ohsu/projects/test","content":[{"attachment":{"contentType":"text/fastq","creation":"2024-08-21T10:53:00+00:00","extension":[{"url":"http://aced-idp.org/fhir/StructureDefinition/md5","valueString":"227f0a5379362d42eaa1814cfc0101b8"},{"url":"http://aced-idp.org/fhir/StructureDefinition/source_path","valueUrl":"file:///home/LabA/specimen_1234_labA.fq.gz"}],"size":5595609484,"title":"specimen_1234_labA.fq.gz","url":"file:///home/LabA/specimen_1234_labA.fq.gz"}}],"date":"2024-08-21T10:53:00+00:00","docStatus":"final","id":"9ae7e542-767f-4b03-a854-7ceed17152cb","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"9ae7e542-767f-4b03-a854-7ceed17152cb"}],"resourceType":"DocumentReference","status":"current","subject":{"reference":"Specimen/60c67a06-ea2d-4d24-9249-418dc77a16a9"}}
+{"_id":"8ae7e542-767f-4b03-a854-7ceed17152cb","_label":"DocumentReference","auth_resource_path":"/programs/ohsu/projects/test-invalid","content":[{"attachment":{"contentType":"text/fastq","creation":"2024-08-21T10:53:00+00:00","extension":[{"url":"http://aced-idp.org/fhir/StructureDefinition/md5","valueString":"227f0a5379362d42eaa1814cfc0101b8"},{"url":"http://aced-idp.org/fhir/StructureDefinition/source_path","valueUrl":"file:///home/LabA/specimen_1234_labA.fq.gz"}],"size":5595609484,"title":"specimen_1234_labA.fq.gz","url":"file:///home/LabA/specimen_1234_labA.fq.gz"}}],"date":"2024-08-21T10:53:00+00:00","docStatus":"final","id":"8ae7e542-767f-4b03-a854-7ceed17152cb","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"8ae7e542-767f-4b03-a854-7ceed17152cb"}],"resourceType":"DocumentReference","status":"current","subject":{"reference":"Specimen/50c67a06-ea2d-4d24-9249-418dc77a16a9"}}
