+++ tests/fixtures/combio-examples-grip/Observation.in.edge.json
@@ -1,14 +1,14 @@
-{"id":"f043f4b2-6eed-5431-aed9-f875390799c2","label":"focus_DocumentReference","from":"cec32723-9ede-5f24-ba63-63cb8c6a02cf","to":"9ae7e542-767f-4b03-a854-7ceed17152cb"}
-{"id":"0130767a-71ff-5677-aeb9-b1ce9f37f25c","label":"specimen_Specimen","from":"cec32723-9ede-5f24-ba63-63cb8c6a02cf","to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id":"7471e5a6-d026-5311-b79a-0a3f7b9e9893","label":"subject_Patient","from":"cec32723-9ede-5f24-ba63-63cb8c6a02cf","to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
-{"id":"4d3335be-7333-5c61-a534-57e2148f3df0","label":"focus_Specimen","from":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061","to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id":"e2867c6d-db7e-5d6e-87d8-b1c293f5b47e","label":"subject_Patient","from":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061","to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
-{"id":"2bc87039-7ec6-539a-b73b-158fbb4ea778","label":"focus_DocumentReference","from":"21f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"9ae7e542-767f-4b03-a854-7ceed17152cb"}
-{"id":"72ec2c73-7dd7-5fc7-97f1-cc4a53707159","label":"specimen_Specimen","from":"21f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id":"a6268b43-7100-5663-bbbb-60dd736873a0","label":"subject_Patient","from":"21f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
-{"id":"1bc87039-7ec6-539a-b73b-158fbb4ea778","label":"focus_DocumentReference","from":"01f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"8ae7e542-767f-4b03-a854-7ceed17152cb"}
-{"id":"62ec2c73-7dd7-5fc7-97f1-cc4a53707159","label":"specimen_Specimen","from":"01f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"50c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id":"z6268b43-7100-5663-bbbb-60dd736873a0","label":"subject_Patient","from":"01f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}
-{"id":"0bc87039-7ec6-539a-b73b-158fbb4ea778","label":"focus_DocumentReference","from":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"8ae7e542-767f-4b03-a854-7ceed17152cb"}
-{"id":"52ec2c73-7dd7-5fc7-97f1-cc4a53707159","label":"specimen_Specimen","from":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"50c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id":"y6268b43-7100-5663-bbbb-60dd736873a0","label":"subject_Patient","from":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f","to":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"f043f4b2-6eed-5431-aed9-f875390799c2","_label":"focus_DocumentReference","_from":"cec32723-9ede-5f24-ba63-63cb8c6a02cf","_to":"9ae7e542-767f-4b03-a854-7ceed17152cb"}
+{"_id":"0130767a-71ff-5677-aeb9-b1ce9f37f25c","_label":"specimen_Specimen","_from":"cec32723-9ede-5f24-ba63-63cb8c6a02cf","_to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"7471e5a6-d026-5311-b79a-0a3f7b9e9893","_label":"subject_Patient","_from":"cec32723-9ede-5f24-ba63-63cb8c6a02cf","_to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"4d3335be-7333-5c61-a534-57e2148f3df0","_label":"focus_Specimen","_from":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061","_to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"e2867c6d-db7e-5d6e-87d8-b1c293f5b47e","_label":"subject_Patient","_from":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061","_to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"2bc87039-7ec6-539a-b73b-158fbb4ea778","_label":"focus_DocumentReference","_from":"21f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"9ae7e542-767f-4b03-a854-7ceed17152cb"}
+{"_id":"72ec2c73-7dd7-5fc7-97f1-cc4a53707159","_label":"specimen_Specimen","_from":"21f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"a6268b43-7100-5663-bbbb-60dd736873a0","_label":"subject_Patient","_from":"21f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"1bc87039-7ec6-539a-b73b-158fbb4ea778","_label":"focus_DocumentReference","_from":"01f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"8ae7e542-767f-4b03-a854-7ceed17152cb"}
+{"_id":"62ec2c73-7dd7-5fc7-97f1-cc4a53707159","_label":"specimen_Specimen","_from":"01f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"50c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"z6268b43-7100-5663-bbbb-60dd736873a0","_label":"subject_Patient","_from":"01f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"0bc87039-7ec6-539a-b73b-158fbb4ea778","_label":"focus_DocumentReference","_from":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"8ae7e542-767f-4b03-a854-7ceed17152cb"}
+{"_id":"52ec2c73-7dd7-5fc7-97f1-cc4a53707159","_label":"specimen_Specimen","_from":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"50c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"y6268b43-7100-5663-bbbb-60dd736873a0","_label":"subject_Patient","_from":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f","_to":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}
