+++ tests/fixtures/combio-examples-grip/Observation.out.edge.json
@@ -1,11 +1,11 @@
-{"id":"b51c4080-6c87-548e-ae5a-27cff55ab422","label":"focus_observation","from":"9ae7e542-767f-4b03-a854-7ceed17152cb","to":"cec32723-9ede-5f24-ba63-63cb8c6a02cf"}
-{"id":"63c39ac6-a32d-5ec6-9a25-fbb37b035300","label":"specimen_observation","from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","to":"cec32723-9ede-5f24-ba63-63cb8c6a02cf"}
-{"id":"2201f601-22b6-514f-9059-ffb17ec11961","label":"subject_observation","from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"cec32723-9ede-5f24-ba63-63cb8c6a02cf"}
-{"id":"6d4b193a-4a6d-5a0e-904f-933dd6fe1d48","label":"focus_observation","from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","to":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061"}
-{"id":"872b44a7-4bf6-5169-a861-d5b66328a474","label":"subject_observation","from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061"}
-{"id":"4579f383-75d0-5cc5-9c5b-bb4e920225bc","label":"focus_observation","from":"9ae7e542-767f-4b03-a854-7ceed17152cb","to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
-{"id":"dcdd2268-d0e8-5d20-81cd-dd341e0d23cf","label":"specimen_observation","from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
-{"id":"7b701827-a13f-5fe9-a700-54ffb651b955","label":"subject_observation","from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
-{"id":"3579f383-75d0-5cc5-9c5b-bb4e920225bc","label":"focus_observation","from":"8ae7e542-767f-4b03-a854-7ceed17152cb","to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
-{"id":"ccdd2268-d0e8-5d20-81cd-dd341e0d23cf","label":"specimen_observation","from":"50c67a06-ea2d-4d24-9249-418dc77a16a9","to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
-{"id":"6b701827-a13f-5fe9-a700-54ffb651b955","label":"subject_observation","from":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
+{"_id":"b51c4080-6c87-548e-ae5a-27cff55ab422","_label":"focus_observation","_from":"9ae7e542-767f-4b03-a854-7ceed17152cb","_to":"cec32723-9ede-5f24-ba63-63cb8c6a02cf"}
+{"_id":"63c39ac6-a32d-5ec6-9a25-fbb37b035300","_label":"specimen_observation","_from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","_to":"cec32723-9ede-5f24-ba63-63cb8c6a02cf"}
+{"_id":"2201f601-22b6-514f-9059-ffb17ec11961","_label":"subject_observation","_from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"cec32723-9ede-5f24-ba63-63cb8c6a02cf"}
+{"_id":"6d4b193a-4a6d-5a0e-904f-933dd6fe1d48","_label":"focus_observation","_from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","_to":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061"}
+{"_id":"872b44a7-4bf6-5169-a861-d5b66328a474","_label":"subject_observation","_from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"4e3c6b59-b1fd-5c26-a611-da4cde9fd061"}
+{"_id":"4579f383-75d0-5cc5-9c5b-bb4e920225bc","_label":"focus_observation","_from":"9ae7e542-767f-4b03-a854-7ceed17152cb","_to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
+{"_id":"dcdd2268-d0e8-5d20-81cd-dd341e0d23cf","_label":"specimen_observation","_from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","_to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
+{"_id":"7b701827-a13f-5fe9-a700-54ffb651b955","_label":"subject_observation","_from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
+{"_id":"3579f383-75d0-5cc5-9c5b-bb4e920225bc","_label":"focus_observation","_from":"8ae7e542-767f-4b03-a854-7ceed17152cb","_to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
+{"_id":"ccdd2268-d0e8-5d20-81cd-dd341e0d23cf","_label":"specimen_observation","_from":"50c67a06-ea2d-4d24-9249-418dc77a16a9","_to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
+{"_id":"6b701827-a13f-5fe9-a700-54ffb651b955","_label":"subject_observation","_from":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"11f3411d-89a4-4bcc-9ce7-b76edb1c745f"}
