+++ tests/fixtures/combio-examples-grip/Organization.vertex.json
@@ -1,1 +1,1 @@
-{"id":"89c8dc4c-2d9c-48c7-8862-241a49a78f14","label":"Organization","data":{"auth_resource_path":"/programs/ohsu/projects/test","id":"89c8dc4c-2d9c-48c7-8862-241a49a78f14","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"LabA_ORGANIZATION"}],"resourceType":"Organization","type":[{"coding":[{"code":"prov","display":"Healthcare Provider","system":"http://terminology.hl7.org/CodeSystem/organization-type"}],"text":"An organization that provides healthcare services."},{"coding":[{"code":"edu","display":"Educational Institute","system":"http://terminology.hl7.org/CodeSystem/organization-type"}],"text":"An educational institution that provides education or research facilities."}]}
+{"_id":"89c8dc4c-2d9c-48c7-8862-241a49a78f14","_label":"Organization","auth_resource_path":"/programs/ohsu/projects/test","id":"89c8dc4c-2d9c-48c7-8862-241a49a78f14","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"LabA_ORGANIZATION"}],"resourceType":"Organization","type":[{"coding":[{"code":"prov","display":"Healthcare Provider","system":"http://terminology.hl7.org/CodeSystem/organization-type"}],"text":"An organization that provides healthcare services."},{"coding":[{"code":"edu","display":"Educational Institute","system":"http://terminology.hl7.org/CodeSystem/organization-type"}],"text":"An educational institution that provides education or research facilities."}]}
