+++ tests/fixtures/combio-examples-grip/Patient.vertex.json
@@ -1,1 +1,1 @@
-{"id":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","label":"Patient","data":{"active":true,"auth_resource_path":"/programs/ohsu/projects/test","id":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"patientX_1234"}],"resourceType":"Patient"}}
+{"_id":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","_label":"Patient","active":true,"auth_resource_path":"/programs/ohsu/projects/test","id":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"patientX_1234"}],"resourceType":"Patient"}
