+++ tests/fixtures/combio-examples-grip/ResearchStudy.vertex.json
@@ -1,1 +1,1 @@
-{"id":"7dacd4d0-3c8e-470b-bf61-103891627d45","label":"ResearchStudy","data":{"auth_resource_path":"/programs/ohsu/projects/test","description":"LabA Clinical Trial Study: FHIR Schema Chorot Integration","id":"7dacd4d0-3c8e-470b-bf61-103891627d45","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"labA"}],"name":"LabA","resourceType":"ResearchStudy","status":"active"}}
+{"_id":"7dacd4d0-3c8e-470b-bf61-103891627d45","_label":"ResearchStudy","auth_resource_path":"/programs/ohsu/projects/test","description":"LabA Clinical Trial Study: FHIR Schema Chorot Integration","id":"7dacd4d0-3c8e-470b-bf61-103891627d45","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"labA"}],"name":"LabA","resourceType":"ResearchStudy","status":"active"}
