+++ tests/fixtures/combio-examples-grip/ResearchSubject.in.edge.json
@@ -1,2 +1,2 @@
-{"id":"9eb4d599-8963-56f0-ab37-991e6dd12b1f","label":"study","from":"2fc448d6-a23b-4b94-974b-c66110164851","to":"7dacd4d0-3c8e-470b-bf61-103891627d45"}
-{"id":"50aa2059-b71e-5d48-a1fe-6e6465257133","label":"subject_Patient","from":"2fc448d6-a23b-4b94-974b-c66110164851","to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"9eb4d599-8963-56f0-ab37-991e6dd12b1f","_label":"study","_from":"2fc448d6-a23b-4b94-974b-c66110164851","_to":"7dacd4d0-3c8e-470b-bf61-103891627d45"}
+{"_id":"50aa2059-b71e-5d48-a1fe-6e6465257133","_label":"subject_Patient","_from":"2fc448d6-a23b-4b94-974b-c66110164851","_to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
