+++ tests/fixtures/combio-examples-grip/ResearchSubject.out.edge.json
@@ -1,2 +1,2 @@
-{"id":"3fdd2f49-409b-5677-9687-71d1c3d8c63f","label":"research_subject","from":"7dacd4d0-3c8e-470b-bf61-103891627d45","to":"2fc448d6-a23b-4b94-974b-c66110164851"}
-{"id":"b633871f-f369-507c-8b93-21dc1eb1eb63","label":"research_subject","from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"2fc448d6-a23b-4b94-974b-c66110164851"}
+{"_id":"3fdd2f49-409b-5677-9687-71d1c3d8c63f","_label":"research_subject","_from":"7dacd4d0-3c8e-470b-bf61-103891627d45","_to":"2fc448d6-a23b-4b94-974b-c66110164851"}
+{"_id":"b633871f-f369-507c-8b93-21dc1eb1eb63","_label":"research_subject","_from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"2fc448d6-a23b-4b94-974b-c66110164851"}
