+++ tests/fixtures/combio-examples-grip/ResearchSubject.vertex.json
@@ -1,1 +1,1 @@
-{"id":"2fc448d6-a23b-4b94-974b-c66110164851","label":"ResearchSubject","data":{"auth_resource_path":"/programs/ohsu/projects/test","id":"2fc448d6-a23b-4b94-974b-c66110164851","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"subjectX_1234"}],"resourceType":"ResearchSubject","status":"active","subject":{"reference":"Patient/bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}}}
+{"_id":"2fc448d6-a23b-4b94-974b-c66110164851","_label":"ResearchSubject","auth_resource_path":"/programs/ohsu/projects/test","id":"2fc448d6-a23b-4b94-974b-c66110164851","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"subjectX_1234"}],"resourceType":"ResearchSubject","status":"active","subject":{"reference":"Patient/bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}}
