+++ tests/fixtures/combio-examples-grip/Specimen.in.edge.json
@@ -1,2 +1,2 @@
-{"id":"e02b3a82-40d1-5252-bf63-73db4472306f","label":"subject_Patient","from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
-{"id":"d02b3a82-40d1-5252-bf63-73db4472306f","label":"subject_Patient","from":"50c67a06-ea2d-4d24-9249-418dc77a16a9","to":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"e02b3a82-40d1-5252-bf63-73db4472306f","_label":"subject_Patient","_from":"60c67a06-ea2d-4d24-9249-418dc77a16a9","_to":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}
+{"_id":"d02b3a82-40d1-5252-bf63-73db4472306f","_label":"subject_Patient","_from":"50c67a06-ea2d-4d24-9249-418dc77a16a9","_to":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}
