+++ tests/fixtures/combio-examples-grip/Specimen.out.edge.json
@@ -1,2 +1,2 @@
-{"id":"45ef3f12-11f5-5739-9c86-80711d158653","label":"specimen","from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
-{"id":"35ef3f12-11f5-5739-9c86-80711d158653","label":"specimen","from":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920","to":"50c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"45ef3f12-11f5-5739-9c86-80711d158653","_label":"specimen","_from":"bc4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"60c67a06-ea2d-4d24-9249-418dc77a16a9"}
+{"_id":"35ef3f12-11f5-5739-9c86-80711d158653","_label":"specimen","_from":"ac4e1aa6-cb52-40e9-8f20-594d9c84f920","_to":"50c67a06-ea2d-4d24-9249-418dc77a16a9"}
