+++ tests/fixtures/combio-examples-grip/Specimen.vertex.json
@@ -1,2 +1,2 @@
-{"id":"60c67a06-ea2d-4d24-9249-418dc77a16a9","label":"Specimen","data":{"auth_resource_path":"/programs/ohsu/projects/test","collection":{"bodySite":{"concept":{"coding":[{"code":"76752008","display":"Breast","system":"http://snomed.info/sct"}],"text":"Breast"}},"collector":{"reference":"Organization/89c8dc4c-2d9c-48c7-8862-241a49a78f14"}},"id":"60c67a06-ea2d-4d24-9249-418dc77a16a9","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"specimen_1234_labA"}],"processing":[{"method":{"coding":[{"code":"117032008","display":"Spun specimen (procedure)","system":"http://snomed.info/sct"},{"code":"Double-Spun","display":"Double-Spun","system":"https://my_demo.org/labA"}],"text":"Spun specimen (procedure)"}}],"resourceType":"Specimen","subject":{"reference":"Patient/bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}}}
-{"id":"50c67a06-ea2d-4d24-9249-418dc77a16a9","label":"Specimen","data":{"auth_resource_path":"/programs/ohsu/projects/test-invalid","collection":{"bodySite":{"concept":{"coding":[{"code":"76752008","display":"Breast","system":"http://snomed.info/sct"}],"text":"Breast"}},"collector":{"reference":"Organization/89c8dc4c-2d9c-48c7-8862-241a49a78f14"}},"id":"50c67a06-ea2d-4d24-9249-418dc77a16a9","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"specimen_1234_labA"}],"processing":[{"method":{"coding":[{"code":"117032008","display":"Spun specimen (procedure)","system":"http://snomed.info/sct"},{"code":"Double-Spun","display":"Double-Spun","system":"https://my_demo.org/labA"}],"text":"Spun specimen (procedure)"}}],"resourceType":"Specimen","subject":{"reference":"Patient/ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}}}
+{"_id":"60c67a06-ea2d-4d24-9249-418dc77a16a9","_label":"Specimen","auth_resource_path":"/programs/ohsu/projects/test","collection":{"bodySite":{"concept":{"coding":[{"code":"76752008","display":"Breast","system":"http://snomed.info/sct"}],"text":"Breast"}},"collector":{"reference":"Organization/89c8dc4c-2d9c-48c7-8862-241a49a78f14"}},"id":"60c67a06-ea2d-4d24-9249-418dc77a16a9","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"specimen_1234_labA"}],"processing":[{"method":{"coding":[{"code":"117032008","display":"Spun specimen (procedure)","system":"http://snomed.info/sct"},{"code":"Double-Spun","display":"Double-Spun","system":"https://my_demo.org/labA"}],"text":"Spun specimen (procedure)"}}],"resourceType":"Specimen","subject":{"reference":"Patient/bc4e1aa6-cb52-40e9-8f20-594d9c84f920"}}
+{"_id":"50c67a06-ea2d-4d24-9249-418dc77a16a9","_label":"Specimen","auth_resource_path":"/programs/ohsu/projects/test-invalid","collection":{"bodySite":{"concept":{"coding":[{"code":"76752008","display":"Breast","system":"http://snomed.info/sct"}],"text":"Breast"}},"collector":{"reference":"Organization/89c8dc4c-2d9c-48c7-8862-241a49a78f14"}},"id":"50c67a06-ea2d-4d24-9249-418dc77a16a9","identifier":[{"system":"https://my_demo.org/labA","use":"official","value":"specimen_1234_labA"}],"processing":[{"method":{"coding":[{"code":"117032008","display":"Spun specimen (procedure)","system":"http://snomed.info/sct"},{"code":"Double-Spun","display":"Double-Spun","system":"https://my_demo.org/labA"}],"text":"Spun specimen (procedure)"}}],"resourceType":"Specimen","subject":{"reference":"Patient/ac4e1aa6-cb52-40e9-8f20-594d9c84f920"}}
